Add "partial" (in-progress) value to results_status
Update State.results_status to return a value of 'partial' if
there has been a volunteer developer associated with the state.

While the status will be displayed as "In-progress" on the front
end, we're using the "partial" value to match the options for
the metadata_status field.

Addresses https://github.com/openelections/dashboard/issues/29,
https://github.com/openelections/openelections.github.io/issues/19